# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pets App API",
        "version": "1.0"
    },
    "paths": {
        "/api/v1/animals": {
            "get": {
                "tags": [
                    "Animals"
                ],
                "summary": "Get list of user animals.",
                "operationId": "462221b2b926884dfcb64648e09f3bfe",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AnimalListItemResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Animals"
                ],
                "summary": "Create a new animal.",
                "operationId": "949307ddd485693196318a8379298d0a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "sex",
                                    "birth_date",
                                    "breed_name",
                                    "weight",
                                    "weight_unit"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Fluffy"
                                    },
                                    "sex": {
                                        "type": "string",
                                        "enum": [
                                            "male",
                                            "female"
                                        ],
                                        "example": "male"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "example": "2021-06-22"
                                    },
                                    "animal_type": {
                                        "type": "string",
                                        "example": "995037a6-5811-4ace-b1f7-4667517dd6e0",
                                        "nullable": true
                                    },
                                    "custom_animal_type": {
                                        "type": "string",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "breed": {
                                        "type": "string",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "custom_breed_name": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "breed_name": {
                                        "type": "string"
                                    },
                                    "metis": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "weight": {
                                        "type": "number",
                                        "format": "double",
                                        "example": "2.5"
                                    },
                                    "weight_unit": {
                                        "type": "string",
                                        "example": "kg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnimalFullResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/animals/{animal}": {
            "get": {
                "tags": [
                    "Animals"
                ],
                "summary": "Get animal details.",
                "operationId": "1e532e728869fe040909991c8ff37148",
                "parameters": [
                    {
                        "name": "animal",
                        "in": "path",
                        "required": true,
                        "example": "995037a6-60b3-4055-aa14-3513aa9824ca"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnimalFullResource"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Animals"
                ],
                "summary": "Delete an animal.",
                "operationId": "2afde2e2fa50ce6d51b3204360a384da",
                "parameters": [
                    {
                        "name": "animal",
                        "in": "path",
                        "required": true,
                        "example": "995037a6-60b3-4055-aa14-3513aa9824ca"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful response"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Animals"
                ],
                "summary": "Update existing animal.",
                "operationId": "f3411e8d86c8a711aa343cc20ab3fb72",
                "parameters": [
                    {
                        "name": "animal",
                        "in": "path",
                        "required": true,
                        "example": "995037a6-60b3-4055-aa14-3513aa9824ca"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "sex",
                                    "birth_date",
                                    "breed_name",
                                    "weight",
                                    "weight_unit"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Fluffy"
                                    },
                                    "sex": {
                                        "type": "string",
                                        "enum": [
                                            "male",
                                            "female"
                                        ],
                                        "example": "male"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "example": "2021-06-22"
                                    },
                                    "animal_type": {
                                        "type": "string",
                                        "example": "995037a6-5811-4ace-b1f7-4667517dd6e0",
                                        "nullable": true
                                    },
                                    "custom_animal_type": {
                                        "type": "string",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "breed": {
                                        "type": "string",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "custom_breed_name": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "breed_name": {
                                        "type": "string"
                                    },
                                    "metis": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "weight": {
                                        "type": "number",
                                        "format": "double",
                                        "example": "2.5"
                                    },
                                    "weight_unit": {
                                        "type": "string",
                                        "example": "kg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnimalFullResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/animals/{animal}/avatar": {
            "post": {
                "tags": [
                    "Animals"
                ],
                "summary": "Update user avatar.",
                "operationId": "d5059ce02f6a7f62d79ec71d6910d5f1",
                "parameters": [
                    {
                        "name": "animal",
                        "in": "path",
                        "required": true,
                        "example": "995037a6-60b3-4055-aa14-3513aa9824ca"
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "properties": {
                                            "avatar": {
                                                "description": "Animal avatar",
                                                "type": "string",
                                                "format": "binary"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "thumb": {
                                            "type": "string"
                                        },
                                        "full": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AnimalFullResource": {
                "properties": {
                    "uuid": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string",
                        "example": "Fluffy"
                    },
                    "birth_date": {
                        "type": "string",
                        "example": "2010-10-02"
                    },
                    "type": {
                        "type": "string",
                        "example": "Type"
                    },
                    "breed": {
                        "type": "string",
                        "example": "Breed"
                    },
                    "sex": {
                        "type": "string",
                        "example": "Male"
                    },
                    "weight": {
                        "type": "string",
                        "example": 5
                    },
                    "avatar": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "thumb": {
                                    "type": "string",
                                    "example": null,
                                    "nullable": true
                                },
                                "full": {
                                    "type": "string",
                                    "example": null,
                                    "nullable": true
                                }
                            },
                            "type": "object"
                        }
                    },
                    "activiry": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "notes": {
                                    "type": "integer",
                                    "example": 2
                                },
                                "events": {
                                    "type": "integer",
                                    "example": 12
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "AnimalListItemResource": {
                "properties": {
                    "uuid": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "breed": {
                        "type": "string"
                    },
                    "sex": {
                        "type": "string"
                    },
                    "avatar": {
                        "type": "string",
                        "nullable": true
                    },
                    "activiry": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "notes": {
                                    "type": "integer"
                                },
                                "events": {
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    }
}